Move main goto logic back to ExecuteSubroutine
