+++ rpgJavaInterpreter-core/src/main/kotlin/com/smeup/rpgparser/interpreter/internal_interpreter.kt
@@ -31,8 +31,7 @@ import com.smeup.rpgparser.parsing.parsetreetoast.resolveAndValidate
 import com.smeup.rpgparser.parsing.parsetreetoast.todo
 import com.smeup.rpgparser.utils.ComparisonOperator.*
 import com.smeup.rpgparser.utils.chunkAs
-import com.smeup.rpgparser.utils.getContainingCompilationUnit
-import com.smeup.rpgparser.utils.peekOrNull
+import com.smeup.rpgparser.utils.indexOfTaggedStatement
 import com.smeup.rpgparser.utils.resizeTo
 import com.strumenta.kolasu.model.Position
 import com.strumenta.kolasu.model.ReferenceByName
@@ -484,14 +483,6 @@ open class InternalInterpreter(
         }
     }
 
-    private fun GotoException.indexOfTaggedStatement(statements: List<Statement>) = statements.indexOfTag(tag)
-
-    private fun GotoTopLevelException.indexOfTaggedStatement(statements: List<Statement>) = statements.indexOfTag(tag)
-
-    private fun List<Statement>.indexOfTag(tag: String) = indexOfFirst {
-        it is TagStmt && it.tag.lowercase() == tag.lowercase()
-    }
-
     private fun caseInsensitiveMap(aMap: Map<String, Value>): Map<String, Value> {
         val result = TreeMap<String, Value>(String.CASE_INSENSITIVE_ORDER)
         result.putAll(aMap)
@@ -532,7 +523,7 @@ open class InternalInterpreter(
      * @param unwrappedStatements The unwrapped list of statements. It is up to the caller to ensure statements are unwrapped.
      * @param offset Offset to start the execution from.
      */
-    private fun executeUnwrappedAt(unwrappedStatements: List<Statement>, offset: Int) {
+    override fun executeUnwrappedAt(unwrappedStatements: List<Statement>, offset: Int) {
         /**
          * As we execute composite statements recursively, trying to sequentially execute
          * the unwrapped statement list would result in executing every statement multiple times.
@@ -1277,7 +1268,7 @@ open class InternalInterpreter(
         sourceProducer?.let { openLoggingScope(statement, it) }
 
         val internalExecute = {
-            val executionTime = measureNanoTime { executeInternalLogic(statement) }.nanoseconds
+            val executionTime = measureNanoTime { statement.execute(this) }.nanoseconds
             sourceProducer?.let { closeLoggingScope(statement, programName, sourceProducer, executionTime) }
         }
         if (trace != null) {
@@ -1358,72 +1349,6 @@ open class InternalInterpreter(
                 statement.loggableEntityName,
                 executionTime
             )
-        }
-    }
-
-    /**
-     * Execute internal logic of a single statement without telemetry nor additional context.
-     */
-    private fun executeInternalLogic(statement: Statement) {
-        when (statement) {
-            is ExecuteSubroutine -> {
-                try {
-                    // Setup subroutine context
-                    MainExecutionContext.getSubroutineStack().push(statement.subroutine)
-
-                    // Execute subroutine, we deal with exceptions later
-                    var throwable = kotlin.runCatching {
-                        statement.execute(this)
-                    }.exceptionOrNull()
-
-                    val unwrappedStatements = statement.subroutine.referred!!.stmts.explode(true)
-                    val containingCU = unwrappedStatements.firstOrNull()?.getContainingCompilationUnit()
-
-                    // Recursive deal with goto
-                    var scope: ReferenceByName<Subroutine>? = statement.subroutine
-                    while (throwable is GotoException && statement.subroutine == scope) {
-                        // If is a goto to the end, exit SR
-                        if (throwable.tag.equals(scope.referred!!.tag, true)) {
-                            throwable = null
-                            break
-                        }
-
-                        // If tag is in top level we need to quit subroutine and rollback context to top level
-                        val topLevelStatements = containingCU?.main?.stmts?.explode(true)
-                        topLevelStatements?.let {
-                            val goto = throwable as GotoException
-                            val topLevelIndex = goto.indexOfTaggedStatement(it)
-                            if (0 <= topLevelIndex && topLevelIndex < it.size) {
-                                throw GotoTopLevelException(goto.tag)
-                            }
-                        }
-
-                        // We need to know the statement unwrapped in order to jump directly into a nested tag
-                        val offset = throwable.indexOfTaggedStatement(unwrappedStatements)
-                        require(0 <= offset && offset < unwrappedStatements.size) { "Offset $offset is not valid." }
-                        throwable = kotlin.runCatching {
-                            executeUnwrappedAt(unwrappedStatements, offset)
-                        }.exceptionOrNull()
-                        scope = MainExecutionContext.getSubroutineStack().peekOrNull()
-                    }
-
-                    // Deal with other types of exceptions
-                    throwable?.let {
-                        when (it) {
-                            is LeaveSrException -> {
-                                // Just catch it, do nothing
-                            }
-                            else -> throw it
-                        }
-                    }
-                } catch (e: Exception) {
-                    throw e
-                } finally {
-                    // Cleanup subroutine context
-                    MainExecutionContext.getSubroutineStack().pop()
-                }
-            }
-            else -> statement.execute(this)
         }
     }
 }
